Finished fixing the RefpackCodex
Now the RefpackCodex works as the original code and can encode/decode data in a way that is compatible with the original VC++98 files.

Updated the compression tests to be more readable and have all necessary data for compression/decompression testing.

Updated version numbers as the project was being fixed and tested, yes it took that many attempts.

Updated the dictionary.

Added the tests to the solution.

Signed-off-by: Victor Matia <vmatir@outlook.com>