Removed native imports docs
Added a new target specific to the native imports libraries that doesn't require docs.

This is because the native imports are as close as possible to the original library and you should be able to use the original docs.

If you can't use the original docs with this code, the code is wrong.

Signed-off-by: Victor Matia <vmatir@outlook.com>